Good Progress
Users can now slot cards into slots as well as remove them. Cards
dynamically filter based on the slot selected as well as if it is being
upgraded. Next will add numbers and how they affect the hero stats. Will
also add the ability to upgrade cards and a visual way of showing that
as well as what card slot is selected.